Romfile: Move namespaces from Snowflake.Scraping namespace into Snowflake.Romfile namespace
