+++ Snowflake.API/Romfile/IFileSignature.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Snowflake.Scraping.FileSignatures
+namespace Snowflake.Romfile
 {
     /// <summary>
     /// Represents the ROM file signature
